--- crates/rome_formatter/src/ts/object_type_members/mod.rs
@@ -1,3 +1,0 @@
-mod any_object_type_member;
-mod object_type_member_list;
-mod property_signature;
