--- crates/rome_formatter/src/ts/type/any_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsAnyType;
-
-impl ToFormatElement for TsAnyType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.any_token().format(formatter)
-    }
-}
