--- crates/rome_formatter/src/ts/type/bigint_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsBigintType;
-
-impl ToFormatElement for TsBigintType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.bigint_token().format(formatter)
-    }
-}
