--- crates/rome_formatter/src/ts/type/boolean_literal_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsBooleanLiteralType;
-
-impl ToFormatElement for TsBooleanLiteralType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.literal().format(formatter)
-    }
-}
