--- crates/rome_formatter/src/ts/type/boolean_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsBooleanType;
-
-impl ToFormatElement for TsBooleanType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.boolean_token().format(formatter)
-    }
-}
