--- crates/rome_formatter/src/ts/type/mod.rs
@@ -1,22 +1,0 @@
-mod any_type;
-mod bigint_literal_type;
-mod bigint_type;
-mod boolean_literal_type;
-mod boolean_type;
-mod infer_type;
-mod never_type;
-mod non_primitive_type;
-mod null_literal_type;
-mod number_literal_type;
-mod number_type;
-mod object_type;
-mod referece_type;
-mod string_literal_type;
-mod string_type;
-mod symbol_type;
-mod this_type;
-mod ts_type;
-mod typeof_type;
-mod undefined_type;
-mod unknown_type;
-mod void_type;
