--- crates/rome_formatter/src/ts/type/never_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsNeverType;
-
-impl ToFormatElement for TsNeverType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.never_token().format(formatter)
-    }
-}
