--- crates/rome_formatter/src/ts/type/string_literal_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsStringLiteralType;
-
-impl ToFormatElement for TsStringLiteralType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.literal_token().format(formatter)
-    }
-}
