--- crates/rome_formatter/src/ts/type/string_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsStringType;
-
-impl ToFormatElement for TsStringType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.string_token().format(formatter)
-    }
-}
