--- crates/rome_formatter/src/ts/type/symbol_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsSymbolType;
-
-impl ToFormatElement for TsSymbolType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.symbol_token().format(formatter)
-    }
-}
