--- crates/rome_formatter/src/ts/type/typeof_type.rs
@@ -1,13 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{
-    format_elements, space_token, FormatElement, FormatResult, Formatter, ToFormatElement,
-};
-use rslint_parser::ast::TsTypeofType;
-
-impl ToFormatElement for TsTypeofType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        let r#typeof = self.typeof_token().format(formatter)?;
-        let expression_name = self.expression_name().format(formatter)?;
-        Ok(format_elements![r#typeof, space_token(), expression_name])
-    }
-}
