--- crates/rome_formatter/src/ts/type/void_type.rs
@@ -1,9 +1,0 @@
-use crate::formatter_traits::FormatTokenAndNode;
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::TsVoidType;
-
-impl ToFormatElement for TsVoidType {
-    fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        self.void_token().format(formatter)
-    }
-}
