+++ crates/rome_formatter/src/ts/types/bigint_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsBigintType, AstNode};
+use rslint_parser::ast::TsBigintType;
+
 impl ToFormatElement for TsBigintType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.bigint_token().format(formatter)
     }
 }
