+++ crates/rome_formatter/src/ts/types/boolean_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsBooleanType, AstNode};
+use rslint_parser::ast::TsBooleanType;
+
 impl ToFormatElement for TsBooleanType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.boolean_token().format(formatter)
     }
 }
