+++ crates/rome_formatter/src/ts/types/non_primitive_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsNonPrimitiveType, AstNode};
+use rslint_parser::ast::TsNonPrimitiveType;
+
 impl ToFormatElement for TsNonPrimitiveType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.object_token().format(formatter)
     }
 }
