+++ crates/rome_formatter/src/ts/types/null_literal_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsNullLiteralType, AstNode};
+use rslint_parser::ast::TsNullLiteralType;
+
 impl ToFormatElement for TsNullLiteralType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.literal_token().format(formatter)
     }
 }
