+++ crates/rome_formatter/src/ts/types/number_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsNumberType, AstNode};
+use rslint_parser::ast::TsNumberType;
+
 impl ToFormatElement for TsNumberType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.number_token().format(formatter)
     }
 }
