+++ crates/rome_formatter/src/ts/types/string_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsStringType, AstNode};
+use rslint_parser::ast::TsStringType;
+
 impl ToFormatElement for TsStringType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.string_token().format(formatter)
     }
 }
