+++ crates/rome_formatter/src/ts/types/symbol_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsSymbolType, AstNode};
+use rslint_parser::ast::TsSymbolType;
+
 impl ToFormatElement for TsSymbolType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.symbol_token().format(formatter)
     }
 }
