+++ crates/rome_formatter/src/ts/types/typeof_type.rs
@@ -1,7 +1,13 @@
-use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsTypeofType, AstNode};
+use crate::formatter_traits::FormatTokenAndNode;
+use crate::{
+    format_elements, space_token, FormatElement, FormatResult, Formatter, ToFormatElement,
+};
+use rslint_parser::ast::TsTypeofType;
+
 impl ToFormatElement for TsTypeofType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        let r#typeof = self.typeof_token().format(formatter)?;
+        let expression_name = self.expression_name().format(formatter)?;
+        Ok(format_elements![r#typeof, space_token(), expression_name])
     }
 }
