+++ crates/rome_formatter/src/ts/types/undefined_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsUndefinedType, AstNode};
+use rslint_parser::ast::TsUndefinedType;
+
 impl ToFormatElement for TsUndefinedType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.undefined_token().format(formatter)
     }
 }
