+++ crates/rome_formatter/src/ts/types/unknown_type.rs
@@ -1,7 +1,9 @@
+use crate::formatter_traits::FormatTokenAndNode;
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::{ast::TsUnknownType, AstNode};
+use rslint_parser::ast::TsUnknownType;
+
 impl ToFormatElement for TsUnknownType {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_verbatim(self.syntax()))
+        self.unknown_token().format(formatter)
     }
 }
